+++ backend/src/index.js
@@ -41,6 +41,7 @@ app.use('/api/printers', require('./routes/printers'));
 app.use('/api/push-subscriptions', require('./routes/pushSubscriptions'));
 app.use('/api/admin-push-subscriptions', require('./routes/adminPushSubscriptions'));
 app.use('/api/employee-auth', require('./routes/employeeAuth'));
+app.use('/api/announcements', require('./routes/announcements'));
 
 mongoose
   .connect(process.env.MONGO_URI)